requirements.txt
streamlit==1.57.0
qdrant-client==1.16.1
sentence-transformers==5.4.1
openai==2.33.0
pydantic==2.13.3
pyyaml==6.0.2
pytest==8.3.5
pytest-mock==3.15.1
